Fixing README
GitHub's markdown parser handles code formatting differently than
does the local machine. Rather than try iteratively to fix the
appearance of the class layout draft markdown in the README,
moving that documentation into a new .md file which can be
downloaded and viewed on a local editor such as ReText.